# backend/openapi_utf8.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "FastAPI",
                 "version":  "0.1.0"
             },
    "paths":  {
                  "/":  {
                            "get":  {
                                        "summary":  "Health Check",
                                        "operationId":  "health_check__get",
                                        "responses":  {
                                                          "200":  {
                                                                      "description":  "Successful Response",
                                                                      "content":  "@{application/json=}"
                                                                  }
                                                      }
                                    }
                        },
                  "/analyze/youtube":  {
                                           "post":  {
                                                        "summary":  "Analyze Youtube",
                                                        "operationId":  "analyze_youtube_analyze_youtube_post",
                                                        "requestBody":  {
                                                                            "content":  {
                                                                                            "application/x-www-form-urlencoded":  "@{schema=}"
                                                                                        },
                                                                            "required":  true
                                                                        },
                                                        "responses":  {
                                                                          "200":  {
                                                                                      "description":  "Successful Response",
                                                                                      "content":  "@{application/json=}"
                                                                                  },
                                                                          "422":  {
                                                                                      "description":  "Validation Error",
                                                                                      "content":  "@{application/json=}"
                                                                                  }
                                                                      }
                                                    }
                                       },
                  "/analyze/file":  {
                                        "post":  {
                                                     "summary":  "Analyze File",
                                                     "operationId":  "analyze_file_analyze_file_post",
                                                     "requestBody":  {
                                                                         "content":  {
                                                                                         "multipart/form-data":  "@{schema=}"
                                                                                     },
                                                                         "required":  true
                                                                     },
                                                     "responses":  {
                                                                       "200":  {
                                                                                   "description":  "Successful Response",
                                                                                   "content":  "@{application/json=}"
                                                                               },
                                                                       "422":  {
                                                                                   "description":  "Validation Error",
                                                                                   "content":  "@{application/json=}"
                                                                               }
                                                                   }
                                                 }
                                    }
              },
    "components":  {
                       "schemas":  {
                                       "Body_analyze_file_analyze_file_post":  {
                                                                                   "properties":  {
                                                                                                      "file":  {
                                                                                                                   "type":  "string",
                                                                                                                   "contentMediaType":  "application/octet-stream",
                                                                                                                   "title":  "File"
                                                                                                               },
                                                                                                      "language":  {
                                                                                                                       "type":  "string",
                                                                                                                       "title":  "Language",
                                                                                                                       "default":  "Tamil"
                                                                                                                   },
                                                                                                      "user_id":  {
                                                                                                                      "type":  "string",
                                                                                                                      "title":  "User Id"
                                                                                                                  }
                                                                                                  },
                                                                                   "type":  "object",
                                                                                   "required":  [
                                                                                                    "file",
                                                                                                    "user_id"
                                                                                                ],
                                                                                   "title":  "Body_analyze_file_analyze_file_post"
                                                                               },
                                       "Body_analyze_youtube_analyze_youtube_post":  {
                                                                                         "properties":  {
                                                                                                            "url":  {
                                                                                                                        "type":  "string",
                                                                                                                        "title":  "Url"
                                                                                                                    },
                                                                                                            "language":  {
                                                                                                                             "type":  "string",
                                                                                                                             "title":  "Language",
                                                                                                                             "default":  "Tamil"
                                                                                                                         },
                                                                                                            "user_id":  {
                                                                                                                            "type":  "string",
                                                                                                                            "title":  "User Id"
                                                                                                                        }
                                                                                                        },
                                                                                         "type":  "object",
                                                                                         "required":  [
                                                                                                          "url",
                                                                                                          "user_id"
                                                                                                      ],
                                                                                         "title":  "Body_analyze_youtube_analyze_youtube_post"
                                                                                     },
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  "@{$ref=#/components/schemas/ValidationError}",
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
                                       "ValidationError":  {
                                                               "properties":  {
                                                                                  "loc":  {
                                                                                              "items":  "@{anyOf=System.Object[]}",
                                                                                              "type":  "array",
                                                                                              "title":  "Location"
                                                                                          },
                                                                                  "msg":  {
                                                                                              "type":  "string",
                                                                                              "title":  "Message"
                                                                                          },
                                                                                  "type":  {
                                                                                               "type":  "string",
                                                                                               "title":  "Error Type"
                                                                                           },
                                                                                  "input":  {
                                                                                                "title":  "Input"
                                                                                            },
                                                                                  "ctx":  {
                                                                                              "type":  "object",
                                                                                              "title":  "Context"
                                                                                          }
                                                                              },
                                                               "type":  "object",
                                                               "required":  [
                                                                                "loc",
                                                                                "msg",
                                                                                "type"
                                                                            ],
                                                               "title":  "ValidationError"
                                                           }
                                   }
                   }
}